pyproject.toml
[project]
name = "fine-tune-text-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
